Stage1: Handle 3-button recovery boot.
Using a combination of epicmtd's REC_BOOT_MAGIC method and the additional
help of Mike Kasick's kexec technique, boot a decoy recovery stored in the
recovery partition if the 3-finger recovery boot (bootmode=2) is detected.
The decoy recovery is a minimal bml kernel with the sole purpose of clearing
the bootmode param, setting the REC_BOOT_MAGIC value and immediately rebooting.